Reorganize pages with a forum base and admin base extending off of a central base page class
