+++ js/src/admin/components/AdminPage.js
@@ -1,0 +1,8 @@
+import Page from '../../common/components/Page';
+
+/**
+ * The `AdminPage` component
+ *
+ * @abstract
+ */
+export default class AdminPage extends Page {}
