+++ js/src/forum/components/ForumPage.js
@@ -1,0 +1,8 @@
+import Page from '../../common/components/Page';
+
+/**
+ * The `ForumPage` component
+ *
+ * @abstract
+ */
+export default class ForumPage extends Page {}
