--- js/src/forum/components/Page.js
@@ -1,14 +1,0 @@
-import Page from '../../common/components/Page';
-
-/**
- * The `ForumPage` component
- *
- * @abstract
- */
-export default class ForumPage extends Page {
-  init() {
-    super.init();
-
-    app.modal.close();
-  }
-}
